# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BackgroundTaskLogEntry": {
        "properties": {
          "level": {
            "title": "Level",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "timestamp",
          "level",
          "message"
        ],
        "title": "BackgroundTaskLogEntry",
        "type": "object"
      },
      "BackgroundTaskResponse": {
        "properties": {
          "chapter_number": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Number"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "log_entries": {
            "items": {
              "$ref": "#/components/schemas/BackgroundTaskLogEntry"
            },
            "title": "Log Entries",
            "type": "array"
          },
          "payload": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload"
          },
          "progress": {
            "title": "Progress",
            "type": "integer"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "stream_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Id"
          },
          "stream_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Type"
          },
          "task_type": {
            "title": "Task Type",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "user_id",
          "task_type",
          "title",
          "status",
          "progress",
          "created_at",
          "updated_at"
        ],
        "title": "BackgroundTaskResponse",
        "type": "object"
      },
      "BackgroundTaskSnapshotResponse": {
        "description": "同一数据库快照中的任务列表与可续传游标。",
        "properties": {
          "resume_cursor": {
            "title": "Resume Cursor",
            "type": "integer"
          },
          "schema_version": {
            "const": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "snapshot_revision": {
            "title": "Snapshot Revision",
            "type": "string"
          },
          "stream_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Id"
          },
          "stream_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stream Type"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/BackgroundTaskResponse"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "tasks",
          "snapshot_revision",
          "resume_cursor"
        ],
        "title": "BackgroundTaskSnapshotResponse",
        "type": "object"
      },
...
      "CreationContextPatch": {
        "additionalProperties": false,
        "description": "只更新显式出现的语义位置字段。",
        "properties": {
          "chapter_number": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Number"
          },
          "desk_section": {
            "anyOf": [
              {
                "enum": [
                  "content",
                  "versions",
                  "evaluation"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Desk Section"
          },
          "inspiration_draft": {
            "anyOf": [
              {
                "maxLength": 100000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspiration Draft"
          },
          "inspiration_turn": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspiration Turn"
          },
          "surface": {
            "anyOf": [
              {
                "enum": [
                  "inspiration",
                  "archive",
                  "writing"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          }
        },
        "title": "CreationContextPatch",
        "type": "object"
      },
      "CreationContextRead": {
        "description": "当前用户单个项目的跨设备创作上下文。",
        "properties": {
          "chapter_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Number"
          },
          "desk_section": {
            "anyOf": [
              {
                "enum": [
                  "content",
                  "versions",
                  "evaluation"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Desk Section"
          },
          "inspiration_draft": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspiration Draft"
          },
          "inspiration_turn": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inspiration Turn"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "surface": {
            "anyOf": [
              {
                "enum": [
                  "inspiration",
                  "archive",
                  "writing"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Surface"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "project_id",
          "updated_at"
        ],
        "title": "CreationContextRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "AI Novel Generator API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/creation-contexts": {
      "get": {
        "operationId": "list_creation_contexts_api_creation_contexts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CreationContextRead"
                  },
                  "title": "Response List Creation Contexts Api Creation Contexts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Creation Contexts",
        "tags": [
          "Creation Contexts"
        ]
      }
    },
    "/api/creation-contexts/{project_id}": {
      "patch": {
        "operationId": "patch_creation_context_api_creation_contexts__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreationContextPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreationContextRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Patch Creation Context",
        "tags": [
          "Creation Contexts"
        ]
      }
    },
...
    "/api/tasks": {
      "get": {
        "operationId": "list_background_tasks_api_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BackgroundTaskResponse"
                  },
                  "title": "Response List Background Tasks Api Tasks Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Background Tasks",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/tasks/events": {
      "get": {
        "description": "从 PostgreSQL event log 按 cursor 推送任务变化；Redis 只负责唤醒。",
        "operationId": "stream_background_tasks_api_tasks_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "stream_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stream Type"
            }
          },
          {
            "in": "query",
            "name": "stream_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stream Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Durable task events"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Stream Background Tasks",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/tasks/snapshot": {
      "get": {
        "operationId": "get_background_task_snapshot_api_tasks_snapshot_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "stream_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stream Type"
            }
          },
          {
            "in": "query",
            "name": "stream_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Stream Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackgroundTaskSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Background Task Snapshot",
        "tags": [
          "Tasks"
        ]
      }
    },
    "/api/tasks/{task_id}": {
      "get": {
        "operationId": "get_background_task_api_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackgroundTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Background Task",
        "tags": [
          "Tasks"
        ]
      }
    },
...
  }
}